trying different event signature

+++ Batchable.sol
@@ -31,5 +31,5 @@ struct Batch {
 
     event BatchCreated(Batch batch);
     event SpotReserved(Spot spot);
-    event RefundRequested(uint _batchId, address _purchaserAddress, uint _quantity);
+    event RefundRequested(uint256 _batchId, address _purchaserAddress, uint8 _quantity);
 }
